new functions
New Functions: parse_vectors, get_tables_for_vectors, table_subset_from_vectors, get_series_info_from_vectors, get_classic_vector_format_df
parse vectors works like parse tables, takes a vector or list of vectors and returns it as an integer or list of integers
get tables for vectors returns a dictionary mapping each vector to its table, as well as a list of all unique tables used by the list of vectors
table subset from vectors returns a dictionary with a key for each unique table mapped to a list of the vectors from the input that are in that table
get series info from vectors calls the StatsCan api of the same name
get classic vector format df works like the old CANSIM site. Give it a list of vectors and an optional start reference date and it will return a dataframe with a column for each vector.